fix warning

+++ src/xbridge/bitcoinrpcconnector.cpp
@@ -1299,7 +1299,7 @@ bool eth_sendTransaction(const std::string & rpcip,
         // o.push_back(Pair("value",      "0x9184e72a"));
 
         char buf[64];
-        sprintf(buf, "%ullx", amount);
+        sprintf(buf, "%ullx", static_cast<unsigned int>(amount));
         o.push_back(Pair("value", buf));
 
         // o.push_back(Pair("data",       "0xd46e8dd67c5d32be8d46e8dd67c5d32be8058bb8eb970870f072445675058bb8eb970870f072445675"));
